pyproject.toml
[project]
name = "bigram"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "black>=25.1.0",
    "lxml>=5.4.0",
    "mypy>=1.15.0",
    "pyyaml>=6.0.2",
    "watchdog==3.0.0",
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"

[tool.ruff]
select = ["E", "F", "I"]     # errors, flakes, import tidy-ups
ignore = ["E501"]            # allow long links

[tool.ruff.per-file-ignores]
"tests/*" = ["S101"]         # allow assert in tests

